- TOC Bump for 6.2 - Update copyright year
While not a change, I just wanted to let everyone know I am likely going to be rewriting a good deal of this addon to make it a bit more efficient. I realize it's quite clunky and a bit wasteful resource-wise and I plan to alleviate that in the near future.